GenStageKafka v0.1.0 GenStageKafka

A module that is a GenStage Consumer, and publishes the records received into a Kafka topic

Functions

handle_events(events, from, state)

The GenStage consumer event handler.

The events have to be a list of maps where each map has ‘key’ and ‘value’. Every map that conforms to these requirements are produced to the configured kafka partition

init(kafka_options)

The init function for the GenStage server.

Setup the Kafka produce and indicate that we are a GenStage consumer

start_link(kafka_options)

Starts the GenStageKafka with the options.

The options handled by the module arei (defaults shown).

:hosts - [localhost: 9092] :topic - “no_topic” :partition - 0

Given as a keyword lists, i.e.

KafkaGenStage.start_link( topic: “my_topic”, hosts: [my_broker.example.com: 9092] )